Move selectCasesNotification into data-set-notifications.
