+++ v3/src/components/case-table/color-cell-text-editor.tsx
@@ -5,7 +5,7 @@ import {
 import React, { ChangeEvent, useCallback, useEffect, useRef, useState } from "react"
 import { textEditorClassname } from "react-data-grid"
 import { useDataSetContext } from "../../hooks/use-data-set-context"
-import { selectCasesNotification } from "../../models/data/data-set-utils"
+import { selectCasesNotification } from "../../models/data/data-set-notifications"
 import { parseColor, parseColorToHex } from "../../utilities/color-utils"
 import { t } from "../../utilities/translation/translate"
 import { TRenderEditCellProps } from "./case-table-types"
@@ -60,7 +60,7 @@ export default function ColorCellTextEditor({ row, column, onRowChange, onClose 
     data?.applyModelChange(() => {
       data.setSelectedCases([])
     }, {
-      notifications: () => selectCasesNotification(data)
+      notifications: selectCasesNotification(data)
     })
   }, [data])
